Did some oraganization
